refactor: dto 통일
